Refactored UI project structure and file locations
