+++ CMakeLists.txt
@@ -96,12 +96,12 @@ set(CLIENT_SOURCES
     src/client/state_handlers/state_disconnect.c
     src/client/state_handlers/chat_state/state_chat.c
     src/client/state_handlers/chat_state/chat.c
-    src/client/state_handlers/chat_state/ui.c
-    src/client/state_handlers/chat_state/chat_entries.c
-    src/client/state_handlers/chat_state/ui_windows/header_window.c
-    src/client/state_handlers/chat_state/ui_windows/chat_window.c
-    src/client/state_handlers/chat_state/ui_windows/log_window.c
-    src/client/state_handlers/chat_state/ui_windows/input_window.c
+    src/client/ui/ui.c
+    src/client/ui/chat_entries.c
+    src/client/ui/ui_windows/header_window.c
+    src/client/ui/ui_windows/chat_window.c
+    src/client/ui/ui_windows/log_window.c
+    src/client/ui/ui_windows/input_window.c
 )
 add_executable(client src/client/client.c ${CLIENT_SOURCES})
 target_include_directories(client PRIVATE "${PROJECT_SOURCE_DIR}/include/client")
