+++ include/client/ui/chat_entries.h
