+++ include/client/ui/ui_windows/log_window.h
