+++ src/client/ui/ui_windows/input_window.c
@@ -1,8 +1,8 @@
-#include <ctype.h>
-#include "chat_state/ui.h"
-#include "chat_state/ui_data.h"
-#include "chat_state/ui/chat_window.h"
-#include "chat_state/ui/input_window.h"
+#include "ui/ui.h"
+#include "ui/chat_entries.h"
+#include "ui/ui_windows/chat_window.h"
+#include "ui/ui_windows/input_window.h"
+#include "ui/ui_data.h"
 
 extern UI ui;
 static WINDOW* s_input_window;
