+++ src/client/ui/ui_windows/log_window.c
@@ -1,6 +1,7 @@
-#include "chat_state/ui.h"
-#include "chat_state/ui_data.h"
-#include "chat_state/ui/log_window.h"
+#include "ui/ui.h"
+#include "ui/chat_entries.h"
+#include "ui/ui_windows/log_window.h"
+#include "ui/ui_data.h"
 
 extern UI ui;
 static WINDOW* s_log_window;
